Fix samples and tutorials

(cherry picked from commit 3ce27498eebfc03a1c6df9effbdb9f0b7890184f)

+++ Samples/basic/sdl2/src/SystemInterfaceSDL2.h
@@ -38,7 +38,7 @@ public:
     Rocket::Core::Input::KeyIdentifier TranslateKey(SDL_Keycode sdlkey);
     int TranslateMouseButton(Uint8 button);
 	int GetKeyModifiers();
-	float GetElapsedTime();
+	double GetElapsedTime();
     bool LogMessage(Rocket::Core::Log::Type type, const Rocket::Core::String& message);
 };
 #endif
